Move auth enabled row
