+++ packages/insomnia-app/app/ui/components/editors/auth/ntlm-auth.tsx
@@ -6,7 +6,7 @@ import { useActiveRequest } from '../../../hooks/use-active-request';
 import { selectSettings } from '../../../redux/selectors';
 import { OneLineEditor } from '../../codemirror/one-line-editor';
 import { PasswordEditor } from '../password-editor';
-import { AuthEnabledRow } from './auth-enabled-row';
+import { AuthEnabledRow } from './components/auth-enabled-row';
 
 interface Props {
   isVariableUncovered: boolean;
